% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-blur.R
\name{create_filter_blur}
\alias{create_filter_blur}
\title{Create an SVG blur filter}
\usage{
create_filter_blur(id, std_dev = 3, ...)
}
\arguments{
\item{id}{id to use for filter.}

\item{std_dev}{standard deviation of the gaussian blue. default: 3}

\item{...}{other arguments ignored}
}
\value{
minisvg::SVGElement object representing a filter
}
\description{
Create a filter to apply Gaussian blur to an element.
}
\details{
For more information see the MDN SVG docs for
\href{https://developer.mozilla.org/en-US/docs/Web/SVG/Element/feGaussianBlur}{feGaussianBlur}
}
\examples{
\dontrun{
# Create an SVG document
library(minisvg)
doc   <- minisvg::svg_doc()

# Create the filter and add to the SVG definitions
blur_filter <- create_filter_blur(id = 'myblur')
doc$defs(blur_filter)

# Create a rectangle with the animation
rect  <- stag$rect(
  x      = "10\%",
  y      = "10\%",
  width  = "80\%",
  height = "80\%",
  fill   = "lightblue",
  stroke = 'black',
  filter = blur_filter
)

# Add this rectangle to the document, show the SVG text, then render it
doc$append(rect)
doc
doc$show()
}
}
