% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-drop-shadow.R
\name{create_filter_drop_shadow}
\alias{create_filter_drop_shadow}
\title{Create an SVG drop shadow filter}
\usage{
create_filter_drop_shadow(
  id,
  dx = 10,
  dy = dx,
  std_dev = 3,
  colour = "#bbb",
  ...
)
}
\arguments{
\item{id}{id to use for filter.}

\item{dx, dy}{offset of shadow in x and y directions. default: 10}

\item{std_dev}{standard deviation of the gaussian for the blur. default: 3}

\item{colour}{default: #bbb}

\item{...}{other arguments ignored}
}
\value{
minisvg::SVGElement object representing a filter
}
\description{
This filter adds a drop shadow to the element it is applied to.
}
\details{
The drop shadow effect is a combination of 4 primitive SVG filters:

\enumerate{
\item{Blur the element}
\item{Offset this blurred image}
\item{Fill the blurred image with the given colour using a flood and a composite}
\item{Layer the drop shadow behind the original image}
}


For more information see the MDN SVG docs for
\href{https://developer.mozilla.org/en-US/docs/Web/SVG/Element/feGaussianBlur}{feGaussianBlur}
}
\examples{
\dontrun{
# Create an SVG document
library(minisvg)
doc   <- minisvg::svg_doc()

# Create the filter and add to the SVG definitions
drop_shadow_filter <- create_filter_drop_shadow(id = 'dshadow')
doc$defs(drop_shadow_filter)

# Create a rectangle with the animation
rect  <- stag$rect(
  x      = "10\%",
  y      = "10\%",
  width  = "80\%",
  height = "80\%",
  fill   = "lightblue",
  stroke = 'black',
  filter = drop_shadow_filter
)

# Add this rectangle to the document, show the SVG text, then render it
doc$append(rect)
doc
doc$show()
}
}
