% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-turbulent-displacement.R
\name{create_filter_turbulent_displacement}
\alias{create_filter_turbulent_displacement}
\title{Create an animated SVG turbulent displacement filter}
\usage{
create_filter_turbulent_displacement(
  id,
  scale = 50,
  octaves = 4,
  fps = 8,
  freqx = 0.05,
  freqy = freqx,
  frames = 20,
  ...
)
}
\arguments{
\item{id}{id to use for filter.}

\item{scale}{scale of the displacement}

\item{octaves}{corresponds to \code{feTurbulence - numOctaves} parameter. default: 4}

\item{fps}{frames per second. default: 8}

\item{freqx, freqy}{base frequency in x and y directions. default: freqx = 0.05, freqy = freqx}

\item{frames}{number of frames. default: 20}

\item{...}{other arguments ignored}
}
\value{
minisvg::SVGElement object representing a filter
}
\description{
Use turblence to displace an element.
}
\details{
The turbulent displacement effect is a combination of 2 primitive SVG filters:

\enumerate{
\item{Create turbulence}
\item{Use the turblence as a displacment map for the object}
}


For more information see the MDN SVG docs for
\href{https://developer.mozilla.org/en-US/docs/Web/SVG/Element/feGaussianBlur}{feTurbulence}
and \href{https://developer.mozilla.org/en-US/docs/Web/SVG/Element/feDisplacementMap}{feDisplacementMap}
}
\examples{
\dontrun{
# Create an SVG document
library(minisvg)
doc   <- minisvg::svg_doc()

# Create the filter and add to the SVG definitions
turb_filter <- create_filter_turbulent_displacement(id = 'turb')
doc$defs(turb_filter)

# Create a rectangle with the animation
rect  <- stag$rect(
  x      = "10\%",
  y      = "10\%",
  width  = "80\%",
  height = "80\%",
  fill   = "lightblue",
  stroke = 'black',
  filter = turb_filter
)

# Add this rectangle to the document, show the SVG text, then render it
doc$append(rect)
doc
doc$show()
}
}
