% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svgfilter.R
\docType{package}
\name{svgfilter}
\alias{svgfilter}
\title{svgfilter: A package of parameterised SVG filters built using \code{minisvg}}
\description{
svgfilter: A package of parameterised SVG filters built using \code{minisvg}
}
